dynamic managed transactor config
